Workaround: Stopped tracking CustomBlocks.tres
From now on the CustomBlocks.tres in the assets folder will be untracked. This is because it was causing merge conflicts on each merge request.
The most up to date resource pack will now be stored in files/correctTexturepack, and should be copied to your assets folder and reapplied whenever
it is changed.